Prepare for the new autorollback system by moving some stuff around
* SLOSlackDeploymentProcess is now RollbackSlackDeploymentProcess as this
  class will power both slo and metric-query based autorollbacks
* Adds a py.typed marker so that PaaSTA can actually use our types in
  the near future
* Separate the DeploymentProcess code from the actual SLO logic. This
  will make things cleaner once we also add a bunch of metric querying
  code
